Revert "[NFC][DebugInfo][DWARF] Create new low-level dwarf library (#145081)"
This reverts commit cbf781f0bdf2f680abbe784faedeefd6f84c246e.

Breaks a couple of buildbots.
