+++ llvm/lib/CodeGen/AsmPrinter/CMakeLists.txt
@@ -38,7 +38,6 @@ add_llvm_component_library(LLVMAsmPrinter
   Core
   DebugInfoCodeView
   DebugInfoDWARF
-  DebugInfoDWARFLowLevel
   MC
   MCParser
   Remarks
