+++ llvm/lib/DWARFLinker/Classic/CMakeLists.txt
@@ -19,7 +19,6 @@ add_llvm_component_library(LLVMDWARFLinkerClassic
   CodeGen
   CodeGenTypes
   DebugInfoDWARF
-  DebugInfoDWARFLowLevel
   DWARFLinker
   MC
   Object
