+++ llvm/tools/dsymutil/CMakeLists.txt
@@ -15,7 +15,6 @@ set(LLVM_LINK_COMPONENTS
   DWARFLinkerClassic
   DWARFLinkerParallel
   DebugInfoDWARF
-  DebugInfoDWARFLowLevel
   MC
   Object
   Option
