+++ llvm/tools/llvm-dwarfdump/CMakeLists.txt
@@ -1,6 +1,5 @@
 set(LLVM_LINK_COMPONENTS
   DebugInfoDWARF
-  DebugInfoDWARFLowLevel
   AllTargetsDescs
   AllTargetsInfos
   MC
