+++ llvm/tools/llvm-dwarfutil/CMakeLists.txt
@@ -11,7 +11,6 @@ set(LLVM_LINK_COMPONENTS
   DWARFLinkerClassic
   DWARFLinkerParallel
   DebugInfoDWARF
-  DebugInfoDWARFLowLevel
   MC
   ObjCopy
   Object
