+++ llvm/unittests/DebugInfo/DWARF/CMakeLists.txt
@@ -4,7 +4,6 @@ set(LLVM_LINK_COMPONENTS
   BinaryFormat
   CodeGenTypes
   DebugInfoDWARF
-  DebugInfoDWARFLowLevel
   MC
   Object
   ObjectYAML
