+++ utils/bazel/llvm-project-overlay/bolt/BUILD.bazel
@@ -106,7 +106,6 @@ cc_library(
         "//llvm:DWARFLinker",
         "//llvm:DWP",
         "//llvm:DebugInfoDWARF",
-        "//llvm:DebugInfoDWARFLowLevel",
         "//llvm:Demangle",
         "//llvm:JITLink",
         "//llvm:MC",
@@ -297,7 +296,6 @@ cc_library(
         "//llvm:BinaryFormat",
         "//llvm:CodeGen",
         "//llvm:DebugInfoDWARF",
-        "//llvm:DebugInfoDWARFLowLevel",
         "//llvm:Demangle",
         "//llvm:ExecutionEngine",
         "//llvm:MC",
