+++ utils/bazel/llvm-project-overlay/llvm/BUILD.bazel
@@ -493,7 +493,6 @@ cc_library(
         ":DebugInfo",
         ":DebugInfoCodeView",
         ":DebugInfoDWARF",
-        ":DebugInfoDWARFLowLevel",
         ":DebugInfoPDB",
         ":Demangle",
         ":MC",
@@ -573,23 +572,8 @@ cc_library(
     deps = [
         ":BinaryFormat",
         ":DebugInfo",
-        ":DebugInfoDWARFLowLevel",
         ":MC",
         ":Object",
-        ":Support",
-        ":TargetParser",
-    ],
-)
-
-cc_library(
-    name = "DebugInfoDWARFLowLevel",
-    srcs = glob([
-        "lib/DebugInfo/DWARF/LowLevel/*.cpp",
-    ]),
-    hdrs = glob(["include/llvm/DebugInfo/DWARF/LowLevel/*.h"]),
-    copts = llvm_copts,
-    deps = [
-        ":BinaryFormat",
         ":Support",
         ":TargetParser",
     ],
@@ -1170,7 +1154,6 @@ cc_library(
         ":Core",
         ":DebugInfo",
         ":DebugInfoDWARF",
-        ":DebugInfoDWARFLowLevel",
         ":Demangle",
         ":Object",
         ":Support",
@@ -2098,7 +2081,6 @@ cc_library(
         ":Core",
         ":DebugInfoCodeView",
         ":DebugInfoDWARF",
-        ":DebugInfoDWARFLowLevel",
         ":IPO",
         ":IRPrinter",
         ":Instrumentation",
@@ -3240,7 +3222,6 @@ cc_library(
         ":CodeGenTypes",
         ":DWARFLinkerBase",
         ":DebugInfoDWARF",
-        ":DebugInfoDWARFLowLevel",
         ":MC",
         ":Support",
         ":Target",
@@ -3259,7 +3240,6 @@ cc_library(
         ":BinaryFormat",
         ":CodeGen",
         ":DebugInfoDWARF",
-        ":DebugInfoDWARFLowLevel",
         ":Support",
         ":Target",
     ],
@@ -3560,7 +3540,6 @@ cc_library(
         ":DWARFLinkerParallel",
         ":DebugInfo",
         ":DebugInfoDWARF",
-        ":DebugInfoDWARFLowLevel",
         ":DsymutilTableGen",
         ":MC",
         ":Object",
@@ -3978,7 +3957,6 @@ cc_binary(
         ":BinaryFormat",
         ":DebugInfo",
         ":DebugInfoDWARF",
-        ":DebugInfoDWARFLowLevel",
         ":MC",
         ":Object",
         ":Support",
@@ -4010,7 +3988,6 @@ cc_binary(
         ":DWARFLinker",
         ":DWARFLinkerParallel",
         ":DebugInfoDWARF",
-        ":DebugInfoDWARFLowLevel",
         ":DwarfutilOptionsTableGen",
         ":MC",
         ":ObjCopy",
@@ -4628,7 +4605,6 @@ cc_library(
         ":DebugInfo",
         ":DebugInfoBTF",
         ":DebugInfoDWARF",
-        ":DebugInfoDWARFLowLevel",
         ":Debuginfod",
         ":Demangle",
         ":MC",
